+++ datasource/nosql/transactions/transaction.go
@@ -1,41 +1,50 @@
-package datasource
+package nosql
 
-import "github.com/jinzhu/gorm"
+import (
+	"context"
+
+	"github.com/go-kivik/kivik"
+)
 
 // Transaction interface for gorm
 type Transaction interface {
-	Begin()                         // Begin transaction
-	Commit()                        // Commit transaction
-	Rollback()                      // Rollback transaction
-	DataSource() interface{}        // Get lowlevel db functions
+	Begin()                               // Begin transaction
+	Commit()                              // Commit transaction
+	Rollback()                            // Rollback transaction
+	DataSource(dbName string) interface{} // Get lowlevel db functions
+	Client() *kivik.Client
 	Queries() map[string]string     // Queries loaded from files
 	LookupQuery(name string) string // Get by name loaded from file query
 }
 
 type transaction struct {
 	Transaction
-	db      *gorm.DB
-	tx      *gorm.DB
+	client  *kivik.Client
 	queries map[string]string
 }
 
 func (t *transaction) Begin() {
-	t.tx = t.db.Begin()
-	if t.tx.Error != nil {
-		panic(t.tx.Error)
-	}
+
 }
 
 func (t *transaction) Commit() {
-	t.tx.Commit()
+	//t.tx.Commit()
 }
 
 func (t *transaction) Rollback() {
-	t.tx.Rollback()
+	//t.tx.Rollback()
 }
 
-func (t *transaction) DataSource() interface{} {
-	return t.tx
+func (t *transaction) DataSource(dbName string) interface{} {
+	db, err := t.client.DB(context.TODO(), dbName)
+	if err != nil {
+		return nil
+	}
+	return db
+}
+
+func (t *transaction) Client() *kivik.Client {
+	return t.client
 }
 
 func (t *transaction) Queries() map[string]string {
@@ -52,22 +61,22 @@ type TransactionFactory interface {
 }
 
 type transactionFactory struct {
-	db      *gorm.DB
+	client  *kivik.Client
 	queries map[string]string
 }
 
-func NewTransactionFactory(db *gorm.DB, queries map[string]string) TransactionFactory {
-	return &transactionFactory{db: db, queries: queries}
+func NewTransactionFactory(client *kivik.Client, queries map[string]string) TransactionFactory {
+	return &transactionFactory{client: client, queries: queries}
 }
 
 func (t transactionFactory) Close() {
-	t.db.Close()
+	//t.db.Close()
 }
 
 func (t transactionFactory) BeginNewTransaction() Transaction {
 	tx := new(transaction)
-	tx.db = t.db
 	tx.queries = t.queries
+	tx.client = t.client
 	tx.Begin()
 	return tx
 }